requirements.txt
streamlit==1.32.2
torch
torchvision
torchaudio
Pillow
matplotlib
opencv-python-headless
numpy
torchinfo
git+https://github.com/jacobgil/pytorch-grad-cam.git



